main

+++ app/Http/Controllers/UserPostController.php
@@ -9,33 +9,61 @@ use App\Models\User;
 use App\Models\Saved;
 use App\Models\Topic;
 use App\Models\BlackList;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\ValidationException;
 
 class UserPostController extends Controller
 {
-    public function home() {
+    public function home(Request $request) {
+        $token = Str::random(32);
+
+        $request->session()->put('post_token', $token);
+
         $topics = Topic::get();
-        $posts = Post::select('posts.*','users.name as admin_name','topics.name as topic_name','users.image as profile_image')
-        ->where('admin_id',Auth::user()->id)
+        $posts = Post::select('posts.*', 'users.name as admin_name', 'topics.name as topic_name', 'users.image as profile_image')
+            ->where('admin_id', Auth::user()->id)
+            ->leftJoin('users', 'posts.admin_id', 'users.id')
+            ->leftJoin('topics', 'posts.topic_id', 'topics.id')
+            ->orderBy('updated_at', 'desc')
+            ->get();
+
+        return view('user.post.createPost', compact('topics', 'posts', 'token'));
+    }
+
+
+    //edit page
+    public function editPage(Request $request , $id) {
+        $token = Str::random(32);
+
+        $request->session()->put('post_token', $token);
+        $post = Post::select('posts.*','users.name as admin_name','topics.name as topic_name')
         ->leftJoin('users','posts.admin_id','users.id')
         ->leftJoin('topics','posts.topic_id','topics.id')
-        ->orderBy('updated_at','desc')
-        ->get();
-        // dd($posts->toArray());
-        return view('user.post.createPost',compact('topics','posts'));
+        ->orderBy('created_at','desc')->get();
+        $post = $post->where('id',$id)->first();
+        $topics = Topic::get();
+        return view('user.post.edit',compact('post','topics','token'));
     }
 
     //Create Post
     public function create(Request $request) {
         if (Auth::user()->id==$request->adminId) {
+            $token = $request->input('token');
+            $storedToken = $request->session()->get('post_token');
+
+            // Validate the one-time token
+            if ($token !== $storedToken) {
+                return back()->with(['error' => 'Invalid token.']);
+            }
             $this->postValidationCheck($request);
             $post = $this->postGetData($request);
             if($request->hasFile('postImage')) {
-                $postImageName = uniqid(). '_' . $request->file('postImage')->getClientOriginalName();
-                $post['image'] = $postImageName;
+                $postImageName = md5($request->file('postImage')->getClientOriginalName()). ".jpg";
+                $post['image'] = $postImageName ;
                 $request->file('postImage')->storeAs('public/',$postImageName);
             }
             $currentDate = Carbon::now()->format('Y-m-d');
@@ -47,6 +75,7 @@ class UserPostController extends Controller
                 return back()->with(['error'=>'Maximum posts reached for today']);
             }elseif($postCount < $maxPostsPerDay) {
                 Post::create($post);
+
             return redirect()->route('user#postHome')->with(['message'=>'Post created successfully']);
             }
         }else {
@@ -62,12 +91,30 @@ class UserPostController extends Controller
 
     //validate post data
     private function postValidationCheck($request) {
-        Validator::make($request->all(),[
-            'desc' => 'required|min:10',
+        $validator = Validator::make($request->all(), [
+            'desc' => 'required|min:10|max:10000',
             'topicId' => 'required',
-            'image' => 'mimes:png,jpg,jpeg,JPEG|file'
-        ])->validate();
+            'token' => 'required|in:' . $request->session()->get('post_token'),
+            'image' => 'mimes:png,jpg,jpeg,gif|file'
+        ])->after(function ($validator) use ($request) {
+            $image = $request->file('postImage');
+            if ($image) {
+                $extension = strtolower($image->getClientOriginalExtension());
+                $allowedExtensions = ['png', 'jpg', 'jpeg', 'gif'];
+
+                if (!in_array($extension, $allowedExtensions)) {
+                    $validator->errors()->add('postImage', 'Invalid image file.');
+                }
+            }
+        });
+
+        if ($validator->fails()) {
+            throw new ValidationException($validator);
+        }
+
+        // Validation passed, continue processing the post creation
     }
+
 
     //get data from post input
     private function postGetData($request) {
@@ -79,39 +126,37 @@ class UserPostController extends Controller
         ];
     }
 
-    //edit page
-    public function editPage(Request $request) {
-        $post = Post::select('posts.*','users.name as admin_name','topics.name as topic_name')
-        ->leftJoin('users','posts.admin_id','users.id')
-        ->leftJoin('topics','posts.topic_id','topics.id')
-        ->orderBy('created_at','desc')->get();
-        $post = $post->where('id',$request->id)->first();
-        $topics = Topic::get();
-        // dd($post->toArray(),$topics->toArray());
-        return view('user.post.edit',compact('post','topics'));
-    }
+
 
     //update post
     public function edit(Request $request) {
+
         $post = Post::find($request->postId);
-        // dd(Auth::user()->id);
-        // dd(Auth::user()->id === $post->admin_id);
+
 
         if (Auth::user()->id == $post->admin_id) {
+            $token = $request->input('token');
+
+            $storedToken = $request->session()->get('post_token');
+            // Validate the one-time token
+            if ($token !== $storedToken) {
+                return back()->with(['error' => 'Invalid token.']);
+            }
+
             $this->postValidationCheck($request);
             $data = $this->postGetData($request);
-            //image check
+
             if($request->hasFile('postImage')) {
-                $dbImage = post::select('image')->where('id',$id)->first();
+                $dbImage = post::select('image')->where('id',$request->postId)->first();
                 $dbImage = $dbImage->image;
             if($dbImage!=null) {
                 Storage::delete('public/'.$dbImage);
             }
-            $imageName = uniqid() . '_' . $request->file('postImage')->getClientOriginalName();
-            $data['image'] = $imageName;
-            $request->file('postImage')->storeAs('public/',$imageName);
+            $postImageName = md5($request->file('postImage')->getClientOriginalName()). ".jpg";
+            $post['image'] = $postImageName ;
+            $request->file('postImage')->storeAs('public/',$postImageName);
         }
-        // dd($data);
+
         post::where('id',$data['id'])->update($data);
 
         return redirect()->route('user#postHome')->with(['message'=>'Post edited successfully']);
